+++ packages/react/pagination/src/lib/react-pagination.spec.tsx
@@ -1,0 +1,10 @@
+import { render } from '@testing-library/react';
+
+import ReactPagination from './react-pagination';
+
+describe('ReactPagination', () => {
+  it('should render successfully', () => {
+    const { baseElement } = render(<ReactPagination />);
+    expect(baseElement).toBeTruthy();
+  });
+});